Renamed item repositories
