+++ repository/src/db_diesel/mod.rs
@@ -9,7 +9,7 @@ mod invoice_line;
 mod invoice_line_row;
 mod invoice_row;
 mod item;
-mod item_query;
+mod item_row;
 mod key_value_store;
 mod location;
 mod location_row;
@@ -53,8 +53,8 @@ pub use invoice::*;
 pub use invoice_line::*;
 pub use invoice_line_row::*;
 pub use invoice_row::*;
-pub use item::ItemRepository;
-pub use item_query::*;
+pub use item::*;
+pub use item_row::ItemRowRepository;
 pub use key_value_store::*;
 pub use location::*;
 pub use location_row::LocationRowRepository;
