+++ server/src/sync/integration_tests/stocktake.rs
@@ -1,7 +1,7 @@
 use chrono::NaiveDate;
 use repository::{
     schema::{LocationRow, StocktakeLineRow, StocktakeRow, StocktakeStatus},
-    InvoiceFilter, InvoiceRepository, ItemFilter, ItemQueryRepository, ItemRepository,
+    InvoiceFilter, InvoiceRepository, ItemFilter, ItemRepository, ItemRowRepository,
     LocationRowRepository, StockLineFilter, StockLineRepository, StocktakeLineRowRepository,
     StocktakeRowRepository, StorageConnection,
 };
@@ -29,7 +29,7 @@ impl SyncRecordTester<Vec<FullStocktake>> for StocktakeRecordTester {
             .upsert_one(&location)
             .unwrap();
 
-        let item = ItemQueryRepository::new(connection)
+        let item = ItemRepository::new(connection)
             .query_one(ItemFilter::new())
             .unwrap()
             .unwrap();
@@ -114,7 +114,7 @@ impl SyncRecordTester<Vec<FullStocktake>> for StocktakeRecordTester {
                             .pop()
                             .unwrap()
                             .stock_line_row;
-                        let item = ItemRepository::new(connection)
+                        let item = ItemRowRepository::new(connection)
                             .find_one_by_id(&stock_line.item_id)
                             .unwrap()
                             .unwrap();
