+++ server/src/sync/translation_central/mod.rs
@@ -16,7 +16,7 @@ use repository::{
         CentralSyncBufferRow, ItemRow, MasterListLineRow, MasterListNameJoinRow, MasterListRow,
         NameRow, StoreRow, UnitRow,
     },
-    ItemRepository, MasterListLineRowRepository, MasterListNameJoinRepository,
+    ItemRowRepository, MasterListLineRowRepository, MasterListNameJoinRepository,
     MasterListRowRepository, NameRepository, RepositoryError, StorageConnection,
     StoreRowRepository, TransactionError, UnitRowRepository,
 };
@@ -149,7 +149,7 @@ fn integrate_record(
     match &record {
         IntegrationUpsertRecord::Name(record) => NameRepository::new(con).upsert_one(record),
         IntegrationUpsertRecord::Unit(record) => UnitRowRepository::new(con).upsert_one(record),
-        IntegrationUpsertRecord::Item(record) => ItemRepository::new(con).upsert_one(record),
+        IntegrationUpsertRecord::Item(record) => ItemRowRepository::new(con).upsert_one(record),
         IntegrationUpsertRecord::Store(record) => StoreRowRepository::new(con).upsert_one(record),
         IntegrationUpsertRecord::MasterList(record) => {
             MasterListRowRepository::new(con).upsert_one(record)
