+++ server/src/sync/translation_central/test_data/mod.rs
@@ -11,7 +11,7 @@ use repository::{
         CentralSyncBufferRow, ItemRow, MasterListLineRow, MasterListNameJoinRow, MasterListRow,
         NameRow, StoreRow, UnitRow,
     },
-    ItemRepository, MasterListLineRowRepository, MasterListNameJoinRepository,
+    ItemRowRepository, MasterListLineRowRepository, MasterListNameJoinRepository,
     MasterListRowRepository, NameRepository, RepositoryError, StorageConnection,
     StoreRowRepository, UnitRowRepository,
 };
@@ -88,7 +88,7 @@ pub async fn check_records_against_database(
             }
             TestSyncDataRecord::Item(comparison_record) => {
                 assert_eq!(
-                    ItemRepository::new(&connection)
+                    ItemRowRepository::new(&connection)
                         .find_one_by_id(&record.central_sync_buffer_row.record_id)
                         .unwrap(),
                     comparison_record
